+++ cogs/Fun.py
@@ -13,12 +13,14 @@ class Fun(commands.Cog):
     async def blame(self, ctx, arg1, arg2):
         username = ctx.message.author.name
         await ctx.send(f"{ctx.author.mention} blamed {arg1} for {arg2}", delete_after=30)
-        await ctx.message.delete()
+        if ctx.guild:
+            await ctx.message.delete()
 
     @commands.command(name="wahrheitoderpflicht", aliases=["wop"], help="SFW Wahrheit oder Pflicht")
     async def wahrheitoderpflicht(self, ctx):
         await ctx.send(f"{ctx.author.mention} you have to do the task : {random.choice(commands.Bot.wop_questions)}",delete_after=20)
-        await ctx.message.delete()
+        if ctx.guild:
+            await ctx.message.delete()
 
 
 
@@ -30,7 +32,8 @@ class Fun(commands.Cog):
         else:
             await ctx.send(f"{ctx.author.mention} slapped {user_to_be_slapped.mention}", delete_after=20)
             await ctx.send("https://tenor.com/bd1Da.gif", delete_after=20)
-        await ctx.message.delete()
+        if ctx.guild:
+            await ctx.message.delete()
 
 
     @commands.command(name="lapogusamogus", aliases=["amogus", "lapog", "sus"], hidden=True)
@@ -38,11 +41,13 @@ class Fun(commands.Cog):
         if user_to_sus is not None:
             await ctx.send(f"{user_to_sus.mention} is now a lapogusamogus")
             await ctx.send("https://tenor.com/view/19dollar-fortnite-card-among-us-amogus-sus-red-among-sus-gif-20549014")
-            await ctx.message.delete()
+            if ctx.guild:
+                await ctx.message.delete()
         else:
             await ctx.send(f"{ctx.author.mention} is now a lapogusamogus")
             await ctx.send("https://tenor.com/view/19dollar-fortnite-card-among-us-amogus-sus-red-among-sus-gif-20549014")
-            await ctx.message.delete()
+            if ctx.guild:
+                await ctx.message.delete()
 
 def setup(bot):
     bot.add_cog(Fun(bot))
